Add files via upload
This function takes your favourite picture and converts it into a colormap.
My idea is to do a principal component analysis (PCA) in the RGB space and transform the RGB data of your favourite image into this new coordinate system. After that the function determines the center of mass of the RGB data while walking along the first principal component. After the path in the principal component space is determined, you do a backward transformation and boom! The colormap captures pretty nicely the distribution of colors in your image. I added an extra feature to allow for a separation between discrete and continuous colormaps.
Have fun!